Break up some of the larger opds files
