+++ src/palace/manager/integration/license/opds/opds2/extractor.py
@@ -2,21 +2,11 @@ from __future__ import annotations
 
 from collections.abc import Sequence
 from datetime import date, datetime
-from typing import Any, cast
+from typing import Any
 from urllib.parse import urljoin, urlparse
 
-from celery.canvas import Signature
-from flask_babel import lazy_gettext as _
 from frozendict import frozendict
-from pydantic import ValidationError
-from sqlalchemy.orm import Session
-from typing_extensions import Self, Unpack
-from uritemplate import URITemplate
 
-from palace.manager.api.circulation.base import BaseCirculationAPI
-from palace.manager.api.circulation.exceptions import CannotFulfill
-from palace.manager.api.circulation.fulfillment import RedirectFulfillment
-from palace.manager.core.exceptions import PalaceValueError
 from palace.manager.data_layer.bibliographic import BibliographicData
 from palace.manager.data_layer.circulation import CirculationData
 from palace.manager.data_layer.contributor import ContributorData
@@ -24,369 +14,21 @@ from palace.manager.data_layer.format import FormatData
 from palace.manager.data_layer.identifier import IdentifierData
 from palace.manager.data_layer.link import LinkData
 from palace.manager.data_layer.subject import SubjectData
-from palace.manager.integration.base import integration_settings_load
-from palace.manager.integration.license.opds.opds1 import (
-    BaseOPDSAPI,
-    OPDSImporterLibrarySettings,
-    OPDSImporterSettings,
-)
-from palace.manager.integration.license.opds.requests import (
-    OPDS2AuthType,
-    get_opds_requests,
-)
-from palace.manager.integration.settings import (
-    ConfigurationFormItem,
-    ConfigurationFormItemType,
-    FormField,
-)
 from palace.manager.opds import opds2, rwpm
-from palace.manager.opds.opds2 import AcquisitionObject, PublicationFeedNoValidation
+from palace.manager.opds.opds2 import AcquisitionObject
 from palace.manager.opds.types.link import CompactCollection
-from palace.manager.service.integration_registry.license_providers import (
-    LicenseProvidersRegistry,
-)
-from palace.manager.sqlalchemy.constants import (
-    IdentifierType,
-    LinkRelations,
-    MediaTypes,
-)
+from palace.manager.sqlalchemy.constants import LinkRelations, MediaTypes
 from palace.manager.sqlalchemy.model.classification import Subject
-from palace.manager.sqlalchemy.model.collection import Collection
 from palace.manager.sqlalchemy.model.contributor import Contributor
-from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.edition import Edition
 from palace.manager.sqlalchemy.model.licensing import (
     DeliveryMechanism,
     LicensePool,
-    LicensePoolDeliveryMechanism,
     RightsStatus,
 )
-from palace.manager.sqlalchemy.model.patron import Patron
 from palace.manager.sqlalchemy.model.resource import Hyperlink, Representation
 from palace.manager.util import first_or_default
-from palace.manager.util.http import HTTP
 from palace.manager.util.log import LoggerMixin
-
-
-class OPDS2ImporterSettings(OPDSImporterSettings):
-    custom_accept_header: str = FormField(
-        default="{}, {};q=0.9, */*;q=0.1".format(
-            opds2.PublicationFeed.content_type(), "application/json"
-        ),
-        form=ConfigurationFormItem(
-            label=_("Custom accept header"),
-            description=_(
-                "Some servers expect an accept header to decide which file to send. You can use */* if the server doesn't expect anything."
-            ),
-            type=ConfigurationFormItemType.TEXT,
-            required=False,
-        ),
-    )
-
-    ignored_identifier_types: list[str] = FormField(
-        default=[],
-        form=ConfigurationFormItem(
-            label=_("List of identifiers that will be skipped"),
-            description=_(
-                "Circulation Manager will not be importing publications with identifiers having one of the selected types."
-            ),
-            type=ConfigurationFormItemType.MENU,
-            required=False,
-            options={
-                identifier_type.value: identifier_type.value
-                for identifier_type in IdentifierType
-            },
-            format="narrow",
-        ),
-    )
-
-
-class OPDS2ImporterLibrarySettings(OPDSImporterLibrarySettings):
-    pass
-
-
-class OPDS2API(BaseOPDSAPI):
-    TOKEN_AUTH_CONFIG_KEY = "token_auth_endpoint"
-
-    @classmethod
-    def settings_class(cls) -> type[OPDS2ImporterSettings]:
-        return OPDS2ImporterSettings
-
-    @classmethod
-    def library_settings_class(cls) -> type[OPDS2ImporterLibrarySettings]:
-        return OPDS2ImporterLibrarySettings
-
-    @classmethod
-    def label(cls) -> str:
-        return "OPDS 2.0 Import"
-
-    @classmethod
-    def description(cls) -> str:
-        return "Import books from a publicly-accessible OPDS 2.0 feed."
-
-    def __init__(self, _db: Session, collection: Collection):
-        super().__init__(_db, collection)
-        self.token_auth_configuration: str | None = (
-            collection.integration_configuration.context.get(self.TOKEN_AUTH_CONFIG_KEY)
-        )
-
-    @classmethod
-    def get_authentication_token(
-        cls, patron: Patron, datasource: DataSource, token_auth_url: str
-    ) -> str:
-        """Get the authentication token for a patron"""
-        log = cls.logger()
-
-        patron_id = patron.identifier_to_remote_service(datasource)
-        url = URITemplate(token_auth_url).expand(patron_id=patron_id)
-        response = HTTP.get_with_timeout(url)
-        if response.status_code != 200:
-            log.error(
-                f"Could not authenticate the patron (authorization identifier: '{patron.authorization_identifier}' "
-                f"external identifier: '{patron_id}'): {str(response.content)}"
-            )
-            raise CannotFulfill()
-
-        # The response should be the JWT token, not wrapped in any format like JSON
-        token = response.text
-        if not token:
-            log.error(
-                f"Could not authenticate the patron({patron_id}): {str(response.content)}"
-            )
-            raise CannotFulfill()
-
-        return token
-
-    def fulfill_token_auth(
-        self,
-        patron: Patron,
-        licensepool: LicensePool,
-        fulfillment: RedirectFulfillment,
-    ) -> RedirectFulfillment:
-        templated = URITemplate(fulfillment.content_link)
-        if "authentication_token" not in templated.variable_names:
-            self.log.warning(
-                "No authentication_token variable found in content_link, unable to fulfill via OPDS2 token auth."
-            )
-            return fulfillment
-
-        if not self.token_auth_configuration:
-            self.log.warning(
-                "No token auth configuration found, unable to fulfill via OPDS2 token auth."
-            )
-            return fulfillment
-
-        token = self.get_authentication_token(
-            patron, licensepool.data_source, self.token_auth_configuration
-        )
-        fulfillment.content_link = templated.expand(authentication_token=token)
-        return fulfillment
-
-    def fulfill(
-        self,
-        patron: Patron,
-        pin: str,
-        licensepool: LicensePool,
-        delivery_mechanism: LicensePoolDeliveryMechanism,
-        **kwargs: Unpack[BaseCirculationAPI.FulfillKwargs],
-    ) -> RedirectFulfillment:
-        fulfillment = super().fulfill(
-            patron, pin, licensepool, delivery_mechanism, **kwargs
-        )
-        if self.token_auth_configuration:
-            fulfillment = self.fulfill_token_auth(patron, licensepool, fulfillment)
-        return fulfillment
-
-    @classmethod
-    def import_task(cls, collection_id: int, force: bool = False) -> Signature:
-        from palace.manager.celery.tasks.opds2 import import_collection
-
-        return import_collection.s(collection_id, force=force)
-
-
-class OPDS2Importer(LoggerMixin):
-
-    def __init__(
-        self,
-        *,
-        username: str | None,
-        password: str | None,
-        feed_base_url: str,
-        data_source: str,
-        ignored_identifier_types: list[str],
-        max_retry_count: int,
-        accept_header: str,
-    ) -> None:
-        """
-        Constructor.
-        """
-        self._request = get_opds_requests(
-            (OPDS2AuthType.BASIC if username and password else OPDS2AuthType.NONE),
-            username,
-            password,
-            feed_base_url,
-        )
-
-        self._data_source_name = data_source
-        self._ignored_identifier_types = ignored_identifier_types
-        self._feed_base_url = feed_base_url
-        self._max_retry_count = max_retry_count
-        self._accept_header = accept_header
-
-    @classmethod
-    def from_collection(
-        cls, collection: Collection, registry: LicenseProvidersRegistry
-    ) -> Self:
-        """Create an instance from a Collection."""
-        if not registry.equivalent(collection.protocol, OPDS2API):
-            raise PalaceValueError(
-                f"Collection {collection.name} [id={collection.id} protocol={collection.protocol}] is not a OPDS2 collection."
-            )
-        settings = integration_settings_load(
-            OPDS2API.settings_class(), collection.integration_configuration
-        )
-        return cls(
-            username=settings.username,
-            password=settings.password,
-            feed_base_url=settings.external_account_id,
-            data_source=settings.data_source,
-            ignored_identifier_types=settings.ignored_identifier_types,
-            max_retry_count=settings.max_retry_count,
-            accept_header=settings.custom_accept_header,
-        )
-
-    def get_feed(self, url: str | None) -> PublicationFeedNoValidation:
-        joined_url = urljoin(self._feed_base_url, url)
-        self.log.info(f"Fetching OPDS2 feed page: {joined_url}")
-        return self._request(
-            "GET",
-            joined_url,
-            parser=PublicationFeedNoValidation.model_validate_json,
-            allowed_response_codes=["2xx"],
-            headers={"Accept": self._accept_header},
-            max_retry_count=self._max_retry_count,
-        )
-
-    @classmethod
-    def next_page(cls, feed: PublicationFeedNoValidation) -> str | None:
-        """Get the next page URL from the feed."""
-        next_link = feed.links.get(
-            rel="next", type=PublicationFeedNoValidation.content_type()
-        )
-        if not next_link:
-            return None
-        return next_link.href
-
-    def _is_identifier_allowed(self, identifier: IdentifierData) -> bool:
-        """Check the identifier and return a boolean value indicating whether CM can import it.
-
-        :param identifier: Identifier object
-        :return: Boolean value indicating whether CM can import the identifier
-        """
-        return identifier.type not in self._ignored_identifier_types
-
-    @classmethod
-    def _get_publication(
-        cls,
-        publication: dict[str, Any],
-    ) -> opds2.BasePublication:
-        try:
-            return opds2.Publication.model_validate(publication)
-        except ValidationError as e:
-            raw_identifier = publication.get("metadata", {}).get("identifier")
-            cls.logger().exception(
-                f"Error validating publication (identifier: {raw_identifier}): {e}"
-            )
-            raise
-
-    @classmethod
-    def update_integration_context(
-        cls, feed: PublicationFeedNoValidation, collection: Collection
-    ) -> bool:
-        """Parse the global feed links. Currently only parses the token endpoint link"""
-        links = feed.links
-        token_auth_link = links.get(rel=Hyperlink.TOKEN_AUTH)
-        if token_auth_link is None:
-            return False
-
-        integration = collection.integration_configuration
-        if (
-            integration.context.get(OPDS2API.TOKEN_AUTH_CONFIG_KEY)
-            == token_auth_link.href
-        ):
-            # No change, so we don't need to update the context.
-            return False
-
-        integration.context_update(
-            {OPDS2API.TOKEN_AUTH_CONFIG_KEY: token_auth_link.href}
-        )
-        return True
-
-    def extract_feed_data(
-        self, feed: PublicationFeedNoValidation
-    ) -> list[BibliographicData]:
-        """
-        Turn an OPDS 2.0 feed into lists of BibliographicData and CirculationData objects.
-        """
-        results = []
-        for publication_dict in feed.publications:
-            try:
-                publication = self._get_publication(publication_dict)
-            except ValidationError as e:
-                raw_identifier = publication_dict.get("metadata", {}).get("identifier")
-                raw_title = publication_dict.get("metadata", {}).get("title")
-                self.log.error(
-                    f"Error validating publication (identifier: {raw_identifier}, title: {raw_title}): {e}"
-                )
-                continue
-
-            feed_self_url = feed.links.get(
-                rel=rwpm.LinkRelations.self, raising=True
-            ).href
-            try:
-                publication_bibliographic = Opds2Extractor.extract_publication_data(
-                    publication, self._data_source_name, feed_self_url
-                )
-            except PalaceValueError:
-                self.log.exception(
-                    "Error extracting publication data. Most likely the publications identifier could not be parsed. Skipping publication."
-                )
-                continue
-
-            # We cast here because we know that Opds2Extractor.extract_publication_data always sets
-            # the primary_identifier_data field to an IdentifierData object.
-            # TODO: Maybe we can tighten up the type hint for BibliographicData to reflect this?
-            identifier = cast(
-                IdentifierData, publication_bibliographic.primary_identifier_data
-            )
-            if identifier is None or not self._is_identifier_allowed(identifier):
-                self.log.warning(
-                    f"Publication {identifier} not imported because its identifier type is not allowed: {identifier.type}"
-                )
-                continue
-
-            results.append(publication_bibliographic)
-
-        return results
-
-    @classmethod
-    def is_changed(cls, session: Session, bibliographic: BibliographicData) -> bool:
-        edition = bibliographic.load_edition(session)
-        if not edition:
-            return True
-
-        # If we don't have any information about the last update time, assume we need to update.
-        if edition.updated_at is None or bibliographic.data_source_last_updated is None:
-            return True
-
-        if bibliographic.data_source_last_updated > edition.updated_at:
-            return True
-
-        cls.logger().info(
-            f"Publication {bibliographic.primary_identifier_data} is unchanged. Last updated at "
-            f"{edition.updated_at}, data source last updated at {bibliographic.data_source_last_updated}"
-        )
-        return False
 
 
 class Opds2Extractor(LoggerMixin):
