+++ tests/manager/integration/license/opds/opds2/test_api.py
@@ -1,135 +1,21 @@
-import datetime
-import json
 from collections.abc import Generator
 from unittest.mock import MagicMock, patch
 
 import pytest
-from pydantic import ValidationError
 from requests import Response
 
 from palace.manager.api.circulation.exceptions import CannotFulfill
 from palace.manager.api.circulation.fulfillment import Fulfillment, RedirectFulfillment
 from palace.manager.celery.tasks import opds2 as opds2_celery
-from palace.manager.integration.license.opds import opds2
-from palace.manager.integration.license.opds.opds2 import (
-    OPDS2API,
-    Opds2Extractor,
-    OPDS2Importer,
-)
-from palace.manager.opds.opds2 import PublicationFeedNoValidation
+from palace.manager.integration.license.opds.opds2 import api
+from palace.manager.integration.license.opds.opds2.api import OPDS2API
 from palace.manager.sqlalchemy.model.collection import Collection
-from palace.manager.sqlalchemy.model.contributor import Contributor
 from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.licensing import (
     LicensePool,
     LicensePoolDeliveryMechanism,
 )
 from tests.fixtures.database import DatabaseTransactionFixture
-from tests.fixtures.files import OPDS2FilesFixture
-
-
-class TestOPDS2Importer:
-    def test__get_publication(
-        self,
-        opds2_files_fixture: OPDS2FilesFixture,
-        caplog: pytest.LogCaptureFixture,
-    ) -> None:
-        # Normally _get_publication just turns a publications dict into a Publication model
-        opds2_feed = json.loads(opds2_files_fixture.sample_text("feed.json"))
-        publication_dict = opds2_feed["publications"][0]
-        publication = OPDS2Importer._get_publication(publication_dict)
-        assert publication.metadata.identifier == "urn:isbn:978-3-16-148410-0"
-
-        # However if there is a validation error, it adds a helpful log message
-        # before raising the validation error
-        with pytest.raises(
-            ValidationError, match="3 validation errors for Publication"
-        ):
-            OPDS2Importer._get_publication({})
-
-        assert "3 validation errors for Publication" in caplog.text
-
-    def test_next_page(self, opds2_files_fixture: OPDS2FilesFixture) -> None:
-        # No next links
-        feed = PublicationFeedNoValidation.model_validate_json(
-            opds2_files_fixture.sample_data("feed.json")
-        )
-        assert OPDS2Importer.next_page(feed) is None
-
-        # Feed has next link
-        feed = PublicationFeedNoValidation.model_validate_json(
-            opds2_files_fixture.sample_data("feed2.json")
-        )
-        assert (
-            OPDS2Importer.next_page(feed)
-            == "http://bookshelf-feed-demo.us-east-1.elasticbeanstalk.com/v1/publications?page=2&limit=100"
-        )
-
-
-class TestOpds2Extractor:
-
-    def test__extract_contributor_roles(self) -> None:
-        _extract_contributor_roles = Opds2Extractor._extract_contributor_roles
-
-        # If there are no roles, the function returns the default
-        assert _extract_contributor_roles([], Contributor.Role.AUTHOR) == [
-            Contributor.Role.AUTHOR
-        ]
-
-        # If the role is unknown, the default is used
-        assert _extract_contributor_roles(["invalid"], Contributor.Role.AUTHOR) == [
-            Contributor.Role.AUTHOR
-        ]
-
-        # Roles are not duplicated
-        assert _extract_contributor_roles(
-            [Contributor.Role.AUTHOR, Contributor.Role.AUTHOR], Contributor.Role.AUTHOR
-        ) == [Contributor.Role.AUTHOR]
-        assert _extract_contributor_roles(
-            ["invalid", "invalid"], Contributor.Role.AUTHOR
-        ) == [Contributor.Role.AUTHOR]
-
-        # Role lookup is not case-sensitive
-        assert _extract_contributor_roles(["aUtHoR"], Contributor.Role.ILLUSTRATOR) == [
-            Contributor.Role.AUTHOR
-        ]
-
-        # Roles can be looked up via marc codes
-        assert _extract_contributor_roles(["AUT"], Contributor.Role.ILLUSTRATOR) == [
-            Contributor.Role.AUTHOR
-        ]
-
-    @pytest.mark.parametrize(
-        "published,expected",
-        [
-            pytest.param(
-                datetime.datetime(2015, 9, 29, 17, 0, tzinfo=datetime.timezone.utc),
-                datetime.date(2015, 9, 29),
-                id="datetime with time info",
-            ),
-            pytest.param(
-                datetime.datetime(2015, 9, 29, 0, 0),
-                datetime.date(2015, 9, 29),
-                id="datetime with no time info",
-            ),
-            pytest.param(
-                datetime.date(2015, 9, 29),
-                datetime.date(2015, 9, 29),
-                id="date",
-            ),
-            pytest.param(
-                None,
-                None,
-                id="none",
-            ),
-        ],
-    )
-    def test__extract_published_date(
-        self,
-        published: datetime.datetime | datetime.date | None,
-        expected: datetime.date | None,
-    ) -> None:
-        assert Opds2Extractor._extract_published_date(published) == expected
 
 
 class Opds2ApiFixture:
@@ -173,7 +59,7 @@ class Opds2ApiFixture:
 def opds2_api_fixture(
     db: DatabaseTransactionFixture,
 ) -> Generator[Opds2ApiFixture, None, None]:
-    with patch.object(opds2, "HTTP") as mock_http:
+    with patch.object(api, "HTTP") as mock_http:
         fixture = Opds2ApiFixture(db, mock_http)
         yield fixture
