+++ tests/manager/integration/license/opds/opds2/test_extractor.py
@@ -1,0 +1,72 @@
+import datetime
+
+import pytest
+
+from palace.manager.integration.license.opds.opds2.extractor import Opds2Extractor
+from palace.manager.sqlalchemy.model.contributor import Contributor
+
+
+class TestOpds2Extractor:
+
+    def test__extract_contributor_roles(self) -> None:
+        _extract_contributor_roles = Opds2Extractor._extract_contributor_roles
+
+        # If there are no roles, the function returns the default
+        assert _extract_contributor_roles([], Contributor.Role.AUTHOR) == [
+            Contributor.Role.AUTHOR
+        ]
+
+        # If the role is unknown, the default is used
+        assert _extract_contributor_roles(["invalid"], Contributor.Role.AUTHOR) == [
+            Contributor.Role.AUTHOR
+        ]
+
+        # Roles are not duplicated
+        assert _extract_contributor_roles(
+            [Contributor.Role.AUTHOR, Contributor.Role.AUTHOR], Contributor.Role.AUTHOR
+        ) == [Contributor.Role.AUTHOR]
+        assert _extract_contributor_roles(
+            ["invalid", "invalid"], Contributor.Role.AUTHOR
+        ) == [Contributor.Role.AUTHOR]
+
+        # Role lookup is not case-sensitive
+        assert _extract_contributor_roles(["aUtHoR"], Contributor.Role.ILLUSTRATOR) == [
+            Contributor.Role.AUTHOR
+        ]
+
+        # Roles can be looked up via marc codes
+        assert _extract_contributor_roles(["AUT"], Contributor.Role.ILLUSTRATOR) == [
+            Contributor.Role.AUTHOR
+        ]
+
+    @pytest.mark.parametrize(
+        "published,expected",
+        [
+            pytest.param(
+                datetime.datetime(2015, 9, 29, 17, 0, tzinfo=datetime.timezone.utc),
+                datetime.date(2015, 9, 29),
+                id="datetime with time info",
+            ),
+            pytest.param(
+                datetime.datetime(2015, 9, 29, 0, 0),
+                datetime.date(2015, 9, 29),
+                id="datetime with no time info",
+            ),
+            pytest.param(
+                datetime.date(2015, 9, 29),
+                datetime.date(2015, 9, 29),
+                id="date",
+            ),
+            pytest.param(
+                None,
+                None,
+                id="none",
+            ),
+        ],
+    )
+    def test__extract_published_date(
+        self,
+        published: datetime.datetime | datetime.date | None,
+        expected: datetime.date | None,
+    ) -> None:
+        assert Opds2Extractor._extract_published_date(published) == expected
