+++ tests/manager/scripts/test_opds2_import.py
@@ -4,7 +4,7 @@ import pytest
 
 from palace.manager.celery.tasks import opds2
 from palace.manager.core.exceptions import PalaceValueError
-from palace.manager.integration.license.opds.opds2 import OPDS2API
+from palace.manager.integration.license.opds.opds2.api import OPDS2API
 from palace.manager.scripts.opds2_import import OPDS2ImportScript, OPDS2ReaperScript
 from tests.fixtures.database import DatabaseTransactionFixture
